Remove dependency on HTML tidy
Removed HTML tidy from the the shell script. Not only does this
remove a dependency, it also removes the only asymmetrical part
of the generate/ungenerate process. This was preventing the script
from being properly reversible. Now (and with supporting changes
to the HTML template) the script runs both ways on the same site
and generates reproducible output each time.