Use rsync and conditionals to improve speed
Use rsync instead of deleting/copying everything every time.
This ensures we'll only copy assets that have changed.

Implemented some conditionals for the Markdown processing so that
we only output HTML files which don't already exist or for which
the source file has changed in the last seven days.

These changes should make a big improvement in the speed on
subsequent runs.