Delete files for which the source was removed
Because we're no longer flushing and recreating everything, we need
to check to see if there are any output files that no longer exist
in the source. rsync takes care of this for assets, but we need
a manual implementation for Markdown/HTML. This works both ways,
both for squeezing and unsqueezing.