Support GNU Prolog
Added support for GNU Prolog. Tested everything with both SWI-Prolog
and GNU Prolog, added support to the bash scripts for both
calls, and updated the Prolog source for compatibility with both.
The source was always ISO, but GNU Prolog had some weird behavior
that meant the code wouldn't run as-is. Hopefully this will make
the source code more portable between Prolog implementations.

This remains slightly messy upon unsqueezing as there's no way to
suppress GNU Prolog's compiling output. On squeezing it's easy
enough just to cut all the lines before the doctype or xml elements.
If you really want to use GNU Prolog (or another Prolog that doesn't
let you switch off compiler output), it's up to you to remove the
extra messages when you decompile. Hopefully this won't be often,
so I'm not going to bust my hump to fix it.