Move Markdown to HTML conversion into Prolog
Previously the call to convert Markdown to HTML was in the shell
script, which had the key disadvantage that we needed to separate
and then re-integrate the headers. The logical place to put this
has always been inside Prolog, the problem being that ISO Prolog
doesn't have any way to execute an external program.

I've worked around this limitation by writing an extensible
markdown_to_html predicate which can be implemented for different
Prolog dialects. Included are SWI-Prolog and GNU-Prolog.

The definition of the Markdown command itself can go in your
site.pl, allowing for different Markdown converters and different
arguments on a per-site basis.