Don't sort the RSS articles in Prolog
Apparently the Prolog sort was creating a weird order. I thought
it would key off the first term (i.e. the date), but I guess not.
Seeing as it doesn't seem to matter to RSS readers, which are probably
smart enough to sort these things on their own according to the user's
preferences, there's not much point in fixing it.