Strip HTML entities from the RSS title
HTML entities in the RSS title weren't being escaped, which caused some
strict XML parsers to fail to read it properly. I tried XML escaping
the title, as we already do with the description, but some RSS
readers don't read the title as HTML and so don't display the entities
correctly. Best thing to do is just strip out any HTML entities from
the title and treat it as plain text.

This fix covers the most likely entities to be used, but support
for others can be added easily enough.