Move unsmarten logic into main loop in unsqueeze.sh
Got rid of a faulty loop that unsmartened the punctuation in the
Markdown files. Now that the call to Prolog is in unsqueeze.sh
there isn't any reason for this to be a separate loop, so I moved
the unsmarten code into the main loop. This fixed a bug where some
Markdown files would be left empty by the unsmarten logic.