Don't call Markdown from inside Prolog
This has always bugged me as it's not ISO Prolog and I believe it's
causing other problems, such as with the `wait` command to wait for
child jobs to complete. I've reinstated some of the older functionality
whereby the shell script is the proper glue between utilities. This
will allow us to stick to strict ISO Prolog and get rid of some of
the workarounds for different dialects, although it does mean we
can no longer specify a different Markdown command per-site.