Remove non-ISO Prolog predicates
Removed the dialect-specific non-ISO Prolog predicates. I'm using
the shell to generate the build date for the RSS feed, but there
is no provision in the POSIX `date` spec for formatting dates other
than now, so I've had to write some custom Prolog predicates to
take the dates of the articles and format them as RFC 822 dates,
which is what's required by the RSS spec.

Both the shell script and the Prolog predicates are expecting articles
to have a date in the format YYYY-MM-DD. I may extend this in the future
to include a time and timezone, but I'm not anticipating having to
support any crazy date format.