Implement smart generation so we don't re-generate existing HTML
Implemented an old feature that I got rid of but deserves to make
a more sophisticated comeback. Only generate HTML files for Markdown
sources that are newer than the last generated RSS. This entails
a couple of additional steps:

- We must delete any HTML files for which a source no longer exists.

- We must exclude HTML from the rsync command so they don't get deleted.

- We should include a -f/--force flag to allow the user to specify
  that they would like to force the generation of all the HTML files.

This last point is a "should" because you can always just delete
the contents of your output folder if you want to regenerate everything,
but then you'd be needlessly copying over all your assets again. Better
to make this part of Squeeze's functionality so it can be done intelligently.