Fix hack to wait for all background jobs to complete
I think I finally nailed this bug. The `wait` command wasn't waiting
for background jobs to complete before generating the RSS, which
meant that the most recent article often wouldn't be included in
the feed. I suspect now the problem was a scope issue. The contents
of the while loop are in a different scope, so a `wait` command
outside the loop won't see background jobs started inside the loop.

The fix is to keep track of the index, determine when we're in the
last iteration, and then run `wait`. This lets us remove the ugly
hack that checks for running `sed` or `smartypants` commands.